Add files via upload
Here, I’ll walk you through how to make a rock, paper, and scissors game with Python. In the rock, paper and scissors game our goal is to create a command-line game where a user has the option to choose between rock, paper and scissors and if the user wins the score is added, and at the end when the user finishes the game, the score is shown to the user.